pyproject.toml
[project]
name = "llm-finetuning-domain-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.47.0",
    "datasets>=3.6.0",
    "evaluate>=0.4.5",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "openai>=1.106.0",
    "peft>=0.17.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.56.0",
    "trl>=0.22.2",
    "tyro>=0.9.31",
    "unsloth>=2025.8.10",
    "uvicorn>=0.35.0",
    "xformers>=0.0.32.post2",
]
